add new utils: phonePrefixes & documentTypes
