+++ src/pages/reservations/Reservations.jsx
@@ -1,0 +1,14 @@
+import CustomerForm from "./CustomerForm";
+
+export default function Reservations() {
+  
+
+  return (
+    <div className="row">
+      <fieldset className="col-sm-12 col-md-4">
+        <legend>Datos personales</legend>
+        <CustomerForm/>
+      </fieldset>
+    </div>
+  );
+}
