+++ src/utilies/documentTypes.js
@@ -1,0 +1,20 @@
+export const documentTypes = [
+  "CC",
+  "CE",
+  "PA",
+  "SC",
+  "CD",
+  "TE",
+  "PEP",
+  "AS",
+  "DU",
+  "CCEX",
+  "CEEX",
+  "PAEX",
+  "SCEX",
+  "CDEX",
+  "TEX",
+  "RNEX",
+  "PEPEX",
+  "ASEX",
+];
