+++ src/utilies/phonePrefixes.js
@@ -1,0 +1,18 @@
+export const phonePrefixes = [
+  { country: "Colombia", prefix: "+57" },
+  { country: "Argentina", prefix: "+54" },
+  { country: "Bolivia", prefix: "+591" },
+  { country: "Brasil", prefix: "+55" },
+  { country: "Chile", prefix: "+56" },
+  { country: "Ecuador", prefix: "+593" },
+  { country: "Guyana", prefix: "+592" },
+  { country: "Paraguay", prefix: "+595" },
+  { country: "Peru", prefix: "+51" },
+  { country: "Surinam", prefix: "+597" },
+  { country: "Uruguay", prefix: "+598" },
+  { country: "Venezuela", prefix: "+58" },
+  { country: "Guayana Francesa", prefix: "+594" },
+  { country: "Islas Malvinas", prefix: "+500" },
+  { country: "Islas Georgias del Sur", prefix: "+500" },
+  { country: "Islas Sandwich del Sur", prefix: "+500" },
+];
